trying to fix

+++ core/src/main/kotlin/org/evomaster/core/languagemodel/data/ollama/OllamaModel.kt
@@ -8,7 +8,7 @@ class OllamaModel {
 
     val modified_at: String = ""
 
-    val size: Int = 0
+    val size: Long = 0
 
     val digest: String = ""
